Fix ChallengeResponse result decoding.
ChallengeResponse() returns a json.RawMessage, which means that there is
an extra level of unmarshalling necessary to get a byte sequence that
can be decoded as JWT.

The RawMessage contains a JSON-encoded string. Effectively, it contains
an additional set of double quotes enclosing the actual string. This
cases a problem when trying to decode it as JWT, as that assumes a
string which a base64 encoding of JSON (the double quote is not a valid
base64 character).

Signed-off-by: setrofim <setrofim@gmail.com>